+++ src/Cake.Common/Tools/NuGet/NuGetAliases.cs
@@ -10,6 +10,7 @@ using Cake.Common.Tools.NuGet.Update;
 using Cake.Core;
 using Cake.Core.Annotations;
 using Cake.Core.IO;
+using Cake.Core.IO.NuGet;
 
 namespace Cake.Common.Tools.NuGet
 {
@@ -34,24 +35,24 @@ namespace Cake.Common.Tools.NuGet
         ///                                     Authors                 = new[] {"John Doe"},
         ///                                     Owners                  = new[] {"Contoso"},
         ///                                     Description             = "The description of the package",
-        ///                                     Summary                 = "Excellent summare of what the package does", 
+        ///                                     Summary                 = "Excellent summare of what the package does",
         ///                                     ProjectUrl              = new Uri("https://github.com/SomeUser/TestNuget/"),
         ///                                     IconUrl                 = new Uri("http://cdn.rawgit.com/SomeUser/TestNuget/master/icons/testnuget.png"),
         ///                                     LicenseUrl              = new Uri("https://github.com/SomeUser/TestNuget/blob/master/LICENSE.md"),
         ///                                     Copyright               = "Some company 2015",
         ///                                     ReleaseNotes            = new [] {"Bug fixes", "Issue fixes", "Typos"},
         ///                                     Tags                    = new [] {"Cake", "Script", "Build"},
-        ///                                     RequireLicenseAcceptance= false,        
+        ///                                     RequireLicenseAcceptance= false,
         ///                                     Symbols                 = false,
         ///                                     NoPackageAnalysis       = true,
         ///                                     Files                   = new [] {
-        ///                                                                          new NuSpecContent {Source = "bin/SlackPRTGCommander.dll", Target = "bin"},
+        ///                                                                          new NuSpecContent {Source = "bin/TestNuget.dll", Target = "bin"},
         ///                                                                       },
-        ///                                     BasePath                = "./src/TestNuget/bin/release", 
+        ///                                     BasePath                = "./src/TestNuget/bin/release",
         ///                                     OutputDirectory         = "./nuget"
         ///                                 };
-        ///     
-        ///     NuGetPack("./nuspec/SlackPRTGCommander.nuspec", nuGetPackSettings);
+        ///
+        ///     NuGetPack("./nuspec/TestNuget.nuspec", nuGetPackSettings);
         /// </code>
         /// </example>
         [CakeMethodAlias]
@@ -64,9 +65,58 @@ namespace Cake.Common.Tools.NuGet
                 throw new ArgumentNullException("context");
             }
 
-            var packer = new NuGetPacker(context.FileSystem, context.Environment, 
-                context.ProcessRunner, context.Log, context.Globber, context.GetToolResolver("NuGet"));
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var packer = new NuGetPacker(context.FileSystem, context.Environment, context.ProcessRunner, context.Log, context.Globber, resolver);
             packer.Pack(nuspecFilePath, settings);
+        }
+
+        /// <summary>
+        /// Creates a NuGet package using the specified settings.
+        /// </summary>
+        /// <param name="context">The context.</param>
+        /// <param name="settings">The settings.</param>
+        /// <example>
+        /// <code>
+        ///     var nuGetPackSettings   = new NuGetPackSettings {
+        ///                                     Id                      = "TestNuget",
+        ///                                     Version                 = "0.0.0.1",
+        ///                                     Title                   = "The tile of the package",
+        ///                                     Authors                 = new[] {"John Doe"},
+        ///                                     Owners                  = new[] {"Contoso"},
+        ///                                     Description             = "The description of the package",
+        ///                                     Summary                 = "Excellent summare of what the package does",
+        ///                                     ProjectUrl              = new Uri("https://github.com/SomeUser/TestNuget/"),
+        ///                                     IconUrl                 = new Uri("http://cdn.rawgit.com/SomeUser/TestNuget/master/icons/testnuget.png"),
+        ///                                     LicenseUrl              = new Uri("https://github.com/SomeUser/TestNuget/blob/master/LICENSE.md"),
+        ///                                     Copyright               = "Some company 2015",
+        ///                                     ReleaseNotes            = new [] {"Bug fixes", "Issue fixes", "Typos"},
+        ///                                     Tags                    = new [] {"Cake", "Script", "Build"},
+        ///                                     RequireLicenseAcceptance= false,
+        ///                                     Symbols                 = false,
+        ///                                     NoPackageAnalysis       = true,
+        ///                                     Files                   = new [] {
+        ///                                                                          new NuSpecContent {Source = "bin/TestNuget.dll", Target = "bin"},
+        ///                                                                       },
+        ///                                     BasePath                = "./src/TestNuget/bin/release",
+        ///                                     OutputDirectory         = "./nuget"
+        ///                                 };
+        ///
+        ///     NuGetPack(nuGetPackSettings);
+        /// </code>
+        /// </example>
+        [CakeMethodAlias]
+        [CakeAliasCategory("Pack")]
+        [CakeNamespaceImport("Cake.Common.Tools.NuGet.Pack")]
+        public static void NuGetPack(this ICakeContext context, NuGetPackSettings settings)
+        {
+            if (context == null)
+            {
+                throw new ArgumentNullException("context");
+            }
+
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var packer = new NuGetPacker(context.FileSystem, context.Environment, context.ProcessRunner, context.Log, context.Globber, resolver);
+            packer.Pack(settings);
         }
 
         /// <summary>
@@ -121,7 +171,8 @@ namespace Cake.Common.Tools.NuGet
                 throw new ArgumentNullException("context");
             }
 
-            var runner = new NuGetRestorer(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, context.GetToolResolver("NuGet"));
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var runner = new NuGetRestorer(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, resolver);
             runner.Restore(targetFilePath, settings);
         }
 
@@ -135,12 +186,12 @@ namespace Cake.Common.Tools.NuGet
         /// <code>
         /// // Get the path to the package.
         /// var package = "./nuget/SlackPRTGCommander.0.0.1.nupkg";
-        /// 
+        ///
         /// // Push the package.
         /// NuGetPush(package, new NuGetPushSettings {
         ///     Source = "http://example.com/nugetfeed",
         ///     ApiKey = "4003d786-cc37-4004-bfdf-c4f3e8ef9b3a"
-        /// }); 
+        /// });
         /// </code>
         /// </example>
         [CakeMethodAlias]
@@ -153,7 +204,8 @@ namespace Cake.Common.Tools.NuGet
                 throw new ArgumentNullException("context");
             }
 
-            var packer = new NuGetPusher(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, context.GetToolResolver("NuGet"));
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var packer = new NuGetPusher(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, resolver);
             packer.Push(packageFilePath, settings);
         }
 
@@ -170,7 +222,7 @@ namespace Cake.Common.Tools.NuGet
         ///                 Name = EnvironmentVariable("PUBLIC_FEED_NAME"),
         ///                 Source = EnvironmentVariable("PUBLIC_FEED_SOURCE")
         ///             };
-        /// 
+        ///
         /// NuGetAddSource(
         ///     name:feed.Name,
         ///     source:feed.Source
@@ -201,13 +253,13 @@ namespace Cake.Common.Tools.NuGet
         ///                                 IsSensitiveSource = true,
         ///                                 Verbosity = NuGetVerbosity.Detailed
         ///                             };
-        /// 
+        ///
         /// var feed = new
         ///             {
         ///                 Name = EnvironmentVariable("PRIVATE_FEED_NAME"),
         ///                 Source = EnvironmentVariable("PRIVATE_FEED_SOURCE")
         ///             };
-        /// 
+        ///
         /// NuGetAddSource(
         ///     name:feed.Name,
         ///     source:feed.Source,
@@ -225,7 +277,8 @@ namespace Cake.Common.Tools.NuGet
                 throw new ArgumentNullException("context");
             }
 
-            var runner = new NuGetSources(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, context.GetToolResolver("NuGet"));
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var runner = new NuGetSources(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, resolver);
             runner.AddSource(name, source, settings);
         }
 
@@ -242,7 +295,7 @@ namespace Cake.Common.Tools.NuGet
         ///                 Name = EnvironmentVariable("PRIVATE_FEED_NAME"),
         ///                 Source = EnvironmentVariable("PRIVATE_FEED_SOURCE")
         ///             };
-        /// 
+        ///
         /// NuGetRemoveSource(
         ///    name:feed.Name,
         ///    source:feed.Source
@@ -273,13 +326,13 @@ namespace Cake.Common.Tools.NuGet
         ///                                 IsSensitiveSource = true,
         ///                                 Verbosity = NuGetVerbosity.Detailed
         ///                             };
-        /// 
+        ///
         /// var feed = new
         ///             {
         ///                 Name = EnvironmentVariable("PRIVATE_FEED_NAME"),
         ///                 Source = EnvironmentVariable("PRIVATE_FEED_SOURCE")
         ///             };
-        /// 
+        ///
         /// NuGetRemoveSource(
         ///    name:feed.Name,
         ///    source:feed.Source,
@@ -297,7 +350,8 @@ namespace Cake.Common.Tools.NuGet
                 throw new ArgumentNullException("context");
             }
 
-            var runner = new NuGetSources(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, context.GetToolResolver("NuGet"));
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var runner = new NuGetSources(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, resolver);
             runner.RemoveSource(name, source, settings);
         }
 
@@ -378,7 +432,8 @@ namespace Cake.Common.Tools.NuGet
                 throw new ArgumentNullException("context");
             }
 
-            var runner = new NuGetSources(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, context.GetToolResolver("NuGet"));
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var runner = new NuGetSources(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, resolver);
             return runner.HasSource(source, settings);
         }
 
@@ -425,10 +480,11 @@ namespace Cake.Common.Tools.NuGet
                 throw new ArgumentNullException("context");
             }
 
-            var runner = new NuGetInstaller(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, context.GetToolResolver("NuGet"));
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var runner = new NuGetInstaller(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, resolver);
             runner.Install(packageId, settings);
         }
-        
+
         /// <summary>
         /// Installs NuGet packages using the specified package configuration.
         /// </summary>
@@ -472,7 +528,8 @@ namespace Cake.Common.Tools.NuGet
                 throw new ArgumentNullException("context");
             }
 
-            var runner = new NuGetInstaller(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, context.GetToolResolver("NuGet"));
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var runner = new NuGetInstaller(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, resolver);
             runner.InstallFromConfig(packageConfigPath, settings);
         }
 
@@ -493,7 +550,8 @@ namespace Cake.Common.Tools.NuGet
                 throw new ArgumentNullException("context");
             }
 
-            var runner = new NuGetSetApiKey(context.Log, context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, context.GetToolResolver("NuGet"));
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var runner = new NuGetSetApiKey(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, resolver);
             runner.SetApiKey(apiKey, source, settings);
         }
 
@@ -529,7 +587,8 @@ namespace Cake.Common.Tools.NuGet
                 throw new ArgumentNullException("context");
             }
 
-            var runner = new NuGetSetProxy(context.Log, context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, context.GetToolResolver("NuGet"));
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var runner = new NuGetSetProxy(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, resolver);
             runner.SetProxy(proxy, username, password, settings);
         }
 
@@ -568,7 +627,8 @@ namespace Cake.Common.Tools.NuGet
                 throw new ArgumentNullException("context");
             }
 
-            var runner = new NuGetUpdater(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, context.GetToolResolver("NuGet"));
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var runner = new NuGetUpdater(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, resolver);
             runner.Update(targetFile, new NuGetUpdateSettings());
         }
 
@@ -595,7 +655,8 @@ namespace Cake.Common.Tools.NuGet
                 throw new ArgumentNullException("context");
             }
 
-            var runner = new NuGetUpdater(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, context.GetToolResolver("NuGet"));
+            var resolver = new NuGetToolResolver(context.FileSystem, context.Environment, context.Globber);
+            var runner = new NuGetUpdater(context.FileSystem, context.Environment, context.ProcessRunner, context.Globber, resolver);
             runner.Update(targetFile, settings);
         }
     }