+++ src/Cake.Common/Tools/NuGet/Restore/NuGetRestorer.cs
@@ -1,19 +1,16 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using Cake.Core;
 using Cake.Core.IO;
-using Cake.Core.Utilities;
+using Cake.Core.IO.NuGet;
 
 namespace Cake.Common.Tools.NuGet.Restore
 {
     /// <summary>
     /// The NuGet package restorer used to restore solution packages.
     /// </summary>
-    public sealed class NuGetRestorer : Tool<NuGetRestoreSettings>
+    public sealed class NuGetRestorer : NuGetTool<NuGetRestoreSettings>
     {
         private readonly ICakeEnvironment _environment;
-        private readonly IToolResolver _nugetToolResolver;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="NuGetRestorer"/> class.
@@ -22,13 +19,12 @@ namespace Cake.Common.Tools.NuGet.Restore
         /// <param name="environment">The environment.</param>
         /// <param name="processRunner">The process runner.</param>
         /// <param name="globber">The globber.</param>
-        /// <param name="nugetToolResolver">The NuGet tool resolver</param>
+        /// <param name="resolver">The NuGet tool resolver</param>
         public NuGetRestorer(IFileSystem fileSystem, ICakeEnvironment environment, 
-            IProcessRunner processRunner, IGlobber globber, IToolResolver nugetToolResolver)
-            : base(fileSystem, environment, processRunner, globber)
+            IProcessRunner processRunner, IGlobber globber, INuGetToolResolver resolver)
+            : base(fileSystem, environment, processRunner, globber, resolver)
         {
             _environment = environment;
-            _nugetToolResolver = nugetToolResolver;
         }
 
         /// <summary>
@@ -106,40 +102,6 @@ namespace Cake.Common.Tools.NuGet.Restore
             builder.Append("-NonInteractive");
 
             return builder;
-        }
-
-        /// <summary>
-        /// Gets the name of the tool.
-        /// </summary>
-        /// <returns>The name of the tool.</returns>
-        protected override string GetToolName()
-        {
-            return _nugetToolResolver.Name;
-        }
-
-        /// <summary>
-        /// Gets the possible names of the tool executable.
-        /// </summary>
-        /// <returns>The tool executable name.</returns>
-        protected override IEnumerable<string> GetToolExecutableNames()
-        {
-            return new[] { "NuGet.exe", "nuget.exe" };
-        }
-
-        /// <summary>
-        /// Gets alternative file paths which the tool may exist in
-        /// </summary>
-        /// <param name="settings">The settings.</param>
-        /// <returns>The default tool path.</returns>
-        protected override IEnumerable<FilePath> GetAlternativeToolPaths(NuGetRestoreSettings settings)
-        {
-            var path = _nugetToolResolver.ResolveToolPath();
-            if (path != null)
-            {
-                return new[] { path };
-            }
-
-            return Enumerable.Empty<FilePath>();
         }
     }
 }